+++ src/pages/endpoint-config.py
@@ -2,32 +2,8 @@ import streamlit as st
 from components.init import init
 from components.menu import menu
 from components.confirmation import dialog_confirmation
-
-
-def read_endpoint_list() -> dict[str:str]:
-    """Put in session the content of the data/saved_endpoints file."""
-
-    # Read the file content
-    file = open('../data/saved_endpoints', 'r')
-    content = file.read().strip()
-    file.close()
-
-    # Parse the content into a list of objects
-    endpoint_list = content.split('\n')
-    endpoints = []
-    for raw_endpoint in endpoint_list:
-        # If an error occurs here, it is because the file has the wrong format
-        try:
-            if raw_endpoint.strip() != "":
-                colon_index = raw_endpoint.index(':')
-                name = raw_endpoint[0:colon_index].strip()
-                url = raw_endpoint[colon_index+1:].strip()
-                endpoints.append({'name':name, 'url':url})
-        except Exception:
-            st.error('File "data/saved_endpoints" is wrongly formatted. Correct it then reload the page.')
-
-    # Put the endpoints information in session
-    st.session_state['all_endpoints'] = endpoints
+from tools.sparql_queries import count_graph_triples, insert, delete
+from tools.utils import readable_number, to_snake_case
 
 
 def write_endpoint_list() -> None:
@@ -44,38 +20,81 @@ def write_endpoint_list() -> None:
     file.close()
 
 
-def delete_endpoint(index: int):
-    """Callback function to delete a endpoint from the list"""
-    del st.session_state['endpoint']
+def delete_endpoint(index) -> None:
+    """Delete the specified endpoint."""
+
     del st.session_state['all_endpoints'][index]
     write_endpoint_list()
+    st.rerun()
+
+
+def change_endpoint(index) -> None:
+    """Change the name/URL of a given endpoint"""
+
+    name_key = f"endpoint-config-endpoint-name-{index}"
+    url_key = f"endpoint-config-endpoint-url-{index}"
+
+    st.session_state['all_endpoints'][index]['name'] = st.session_state[name_key]
+    st.session_state['all_endpoints'][index]['url'] = st.session_state[url_key]
+    write_endpoint_list()
+
+
+def delete_graph(graph_uri) -> None:
+    """Delete all statement of a given graph"""
+
+    triple = (graph_uri, '?predicate', '?object')
+    delete([triple], graph=graph_uri)
+    del st.session_state['all_graphs']
+    st.rerun()
+
+
 
 
 ##### Modals #####
 
-@st.dialog('Add a new endpoint')
-def dialog_add_new_endpoint():
+@st.dialog("Add an endpoint")
+def dialog_add_endpoint():
+    """Form to add an endpoint"""
 
-    # User inputs
-    new_name = st.text_input('Endpoint name')
-    new_url = st.text_input('Endpoint URL')
-    btn = st.button('Save')
+    endpoint_name = st.text_input('Endpoint name', value="", placeholder="Write an endpoint name")
+    endpoint_url = st.text_input('Endpoint URL', value="", placeholder="Write an endpoint URL")
+    st.text("")
+    if st.button('Save') and endpoint_name and endpoint_url:
+        # Add to the session, write on disk, and rerun the page
+        st.session_state['all_endpoints'].append({'name': endpoint_name, 'url': endpoint_url})
+        write_endpoint_list()
+        st.rerun()
 
-    # If we have everything filled, and the user clicks on the button:
-    if new_name and new_url and btn:
-        # We need to verify that the endpoint is new (name and URL)
-        endpoint_labels = list(map(lambda endpoint: endpoint['name'], st.session_state['all_endpoints']))
-        endpoint_urls = list(map(lambda endpoint: endpoint['url'], st.session_state['all_endpoints']))
-        if new_name in endpoint_labels:
-            st.error('There is already an endpoint called like that.')
-        elif new_url in endpoint_urls:
-            st.error('There is already an endpoint with this URL')
-        else:
-            # If the endpoint is really new, add to the session, write on disk, and rerun the page
-            st.session_state['all_endpoints'].append({'name': new_name, 'url': new_url})
-            write_endpoint_list()
-            st.rerun()
 
+@st.dialog('Create a graph')
+def dialog_create_graph():
+    """Form to create a graph on the selected endpoint"""
+
+    graph_name = st.text_input('Graph name', value="", placeholder="Write a graph name")
+    graph_comment = st.text_area('Graph comment', value="", placeholder="Write a graph comment (description)")
+    st.text("")
+    if st.button('Save') and graph_name and graph_comment:
+        name = to_snake_case(graph_name)
+        graph_uri = st.session_state['endpoint']['url'] + '/' + name
+
+        # Create the triples
+        triple_name = (graph_uri, 'rdfs:label', f"'{graph_name}'")
+        triple_comment = (graph_uri, 'rdfs:comment', f"'{graph_comment}'")       
+
+        # Insert triples
+        insert([triple_name, triple_comment], graph=graph_uri)
+        del st.session_state['all_graphs']
+
+        st.rerun()
+
+
+
+##### Session initialization #####
+
+if 'endpoint-config-endpoints-list' not in st.session_state:
+    st.session_state['endpoint-config-endpoints-list'] = False
+if 'endpoint-config-graph-list' not in st.session_state:
+    st.session_state['endpoint-config-graph-list'] = False
 
 ##### The page #####
 
@@ -86,53 +105,67 @@ st.title("Endpoint configuration")
 
 st.divider()
 
-st.markdown("### Select your working endpoint")
+# Endpoint list titles and buttons
+col1, col2, col3 = st.columns([5, 2, 2], vertical_alignment='bottom')
+col1.markdown('### Endpoints List')
+if col2.button('Display endpoints'):
+    st.session_state['endpoint-config-endpoints-list'] = True
 
-# -- First part: Known Endpoints
-
-col1, col2 = st.columns([3, 1], vertical_alignment='bottom')
-
-# Here we read and parse the disk, display endpoints, and allow user to choose which one he wants
-read_endpoint_list()
-endpoint_labels = list(map(lambda endpoint: endpoint['name'], st.session_state['all_endpoints']))
-endpoint_urls = list(map(lambda endpoint: endpoint['url'], st.session_state['all_endpoints']))
-
-# If an endpoint is already selected, find it, otherwise just init to Nothing
-if 'endpoint' in st.session_state:
-    selected_index = endpoint_labels.index(st.session_state['endpoint']['name'])
-else:
-    selected_index = None
-
-# If there are endpoints on disk, allow selection, otherwise display a message
-if len(endpoint_labels):
-    selected_label = col1.selectbox('Known endpoints', endpoint_labels, key="selectbox-endpoint", index=selected_index, placeholder="Choose an endpoint")
-else:
-    st.info('No endpoints are saved on disk')
-    selected_label = None
-
-# If there is a selected endpoint, 
-if selected_label :
-    # That is different that the one in session
-    if 'endpoint' not in st.session_state or selected_label != st.session_state['endpoint']['name']:
-        # Update the one in session (or set)
-        selected_index = endpoint_labels.index(selected_label)
-        selected_url = endpoint_urls[selected_index]
-        st.session_state['endpoint'] = {'name': selected_label, 'url': selected_url}
-        # Clear the cache
-        st.cache_data.clear()
-        st.cache_resource.clear()
+if st.session_state['endpoint-config-endpoints-list']:
+    
+    # Endpoint list titles and buttons
+    if col3.button('Hide endpoints'):
+        st.session_state['endpoint-config-endpoints-list'] = False
         st.rerun()
 
-    # Allow the user to remove an endpoint from disk
-    if col2.button('Remove endpoint'):
-        dialog_confirmation(f'This will delete "{selected_label}" from your favorites.', delete_endpoint, index=selected_index)
+    # Display all saved endpoints
+    for i, endpoint in enumerate(st.session_state['all_endpoints']):
+        col1, col2, col3 = st.columns([6, 12, 1], vertical_alignment='bottom')
+        col1.text_input('Name', value=endpoint['name'], key=f"endpoint-config-endpoint-name-{i}", on_change=change_endpoint, kwargs={'index': i})
+        col2.text_input('URL', value=endpoint['url'], key=f"endpoint-config-endpoint-url-{i}", on_change=change_endpoint, kwargs={'index': i})
+
+        # Button to delete a saved endpoint
+        if col3.button(f'🗑️', key=f"endpoint-config-{i}"):
+            dialog_confirmation(f"You are about to delete \"{endpoint['name']}\" endpoint.", delete_endpoint, index=i)
+
+    st.text("")
+    
+    if st.button('Add another endpoint'):
+        dialog_add_endpoint()
 
 
-# -- Second part: Allow user to write a new endpoint
+st.divider()
 
-st.text('') # Spacer
-st.text('') # Spacer
-st.markdown('Your endpoint is not in the list?')
+col1, col2, col3 = st.columns([5, 2, 2], vertical_alignment='bottom')
+col1.markdown('### Endpoint graphs')
+if col2.button('Display graphs'):
+    st.session_state['endpoint-config-graph-list'] = True
 
-if st.button('Add a new endpoint'):
-    dialog_add_new_endpoint()
+if st.session_state['endpoint-config-graph-list']:
+
+    # In case the user did not yet choose an endpoint
+    if 'endpoint' not in st.session_state:
+        st.warning('Please select an endpoint first')
+    else:
+    
+        # Endpoint list titles and buttons
+        if col3.button('Hide graphs'):
+            st.session_state['endpoint-config-graph-list'] = False
+            st.rerun()
+
+
+        st.text("")
+        for i, graph in enumerate(st.session_state['all_graphs']):
+
+            col1, col2, col3, col4 = st.columns([3, 2, 5, 2], vertical_alignment='bottom')
+            col1.markdown(graph['label'])
+            col2.markdown(f"{readable_number(count_graph_triples(graph['uri']))} triples")
+            col3.markdown(graph['comment'])
+            if col4.button('🗑️', key=f"endpoint-config-graph-{i}"):
+                dialog_confirmation(f'You are about to delete the graph "{graph["label"]}".', delete_graph, graph_uri=graph['uri'])
+
+
+        st.text("")
+
+        if st.button('Create a graph'):
+            dialog_create_graph()