+++ src/tools/utils.py
@@ -1,3 +1,6 @@
+import re
+import unicodedata
+
 def ensure_uri(supposed_uri: str) -> str:
     """
     Make sure that the given URI has the correct format.
@@ -34,3 +37,32 @@ def ensure_uri(supposed_uri: str) -> str:
     if not uri.endswith(">"):
         uri = uri + ">"
     return uri
+
+
+def readable_number(number: float) -> str:
+    """Convert the given number into a more readable string"""
+
+    for x in ["", "k", "M", "B"]:
+        if number < 1000.0:
+            return str(round(number, 1)) + x
+        number /= 1000.0
+    raise Exception("This Exception should never happen")
+
+
+
+def to_snake_case(text):
+    """Format the given string into snake-case"""
+
+    # Normalize text to decompose accented characters (e.g., é -> e)
+    normalized_text = unicodedata.normalize("NFKD", text)
+
+    # Remove diacritics (accents) by filtering out non-ASCII characters
+    no_accents_text = "".join([c for c in normalized_text if not unicodedata.combining(c)])
+
+    # Remove punctuation
+    cleaned_text = re.sub(r"[^\w\s]", "", no_accents_text)
+
+    # Replace spaces with underscores and convert to lowercase
+    snake_case_text = re.sub(r"\s+", "_", cleaned_text.strip()).lower()
+
+    return snake_case_text